# docs/openapi.json
{
  "components": {
    "schemas": {
      "Category": {
        "enum": [
          "school",
          "activities",
          "medical",
          "friends",
          "other"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "EventInput": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category",
            "default": "other"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "starts_at",
          "ends_at",
          "member_id"
        ],
        "title": "EventInput",
        "type": "object"
      },
      "EventOutput": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category",
            "default": "other"
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "member_id": {
            "format": "uuid",
            "title": "Member Id",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "starts_at",
          "ends_at",
          "member_id",
          "id"
        ],
        "title": "EventOutput",
        "type": "object"
      },
      "EventUpdate": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Category"
              },
              {
                "type": "null"
              }
            ]
          },
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "starts_at",
          "ends_at"
        ],
        "title": "EventUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MemberInput": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Color",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "color"
        ],
        "title": "MemberInput",
        "type": "object"
      },
      "MemberOutput": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "pattern": "^#[0-9a-fA-F]{6}$",
            "title": "Color",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "photo_data": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Data"
          }
        },
        "required": [
          "name",
          "color",
          "id"
        ],
        "title": "MemberOutput",
        "type": "object"
      },
      "MemberUpdate": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "photo_data": {
            "anyOf": [
              {
                "maxLength": 2800000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Photo Data"
          }
        },
        "required": [
          "name"
        ],
        "title": "MemberUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2AuthorizationCodeBearer": {
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "http://localhost:8080/auth/realms/family/protocol/openid-connect/auth",
            "scopes": {},
            "tokenUrl": "http://localhost:8080/auth/realms/family/protocol/openid-connect/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Family Planner API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "operationId": "list_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "member_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventOutput"
                  },
                  "title": "Response List Events Api V1 Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "List Events",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "add_event_api_v1_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "Add Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "delete": {
        "operationId": "delete_event_api_v1_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "Delete Event",
        "tags": [
          "events"
        ]
      },
      "put": {
        "operationId": "update_event_api_v1_events__event_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "Update Event",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/members": {
      "get": {
        "operationId": "list_members_api_v1_members_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MemberOutput"
                  },
                  "title": "Response List Members Api V1 Members Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "List Members",
        "tags": [
          "members"
        ]
      },
      "post": {
        "operationId": "add_member_api_v1_members_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "Add Member",
        "tags": [
          "members"
        ]
      }
    },
    "/api/v1/members/{member_id}": {
      "put": {
        "operationId": "update_member_api_v1_members__member_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Member Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemberUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberOutput"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ],
        "summary": "Update Member",
        "tags": [
          "members"
        ]
      }
    },
...
  }
}